Join hierarchical permissions for scopable resources
ERICA allows to restrict permissions to certain resources (Study,
Center or Patient). When restricting a role to a specific Center,
we want to allow the roles permissions to be granted to all resources
in the hierarchical structure.

Example:

- `Role 1` grants `read_study`, `manage_center`, `read_patient`.
- `Study 1` => `Center 1` => `Patient 1`, `Patient 2`, `Patient 3`

Case #1: `Role 1` is restricted to `Center 1`

The user should be able to:

- read `Study 1`
- read, update, destroy `Center 1`
- read `Patient 1`, `Patient 2`, `Patient 3`

Case #2: `Role 1` is restricted to `Patient 1`

The user should be able to:

- read `Study 1`
- read, update, destroy `Center 1`
- read `Patient 1`

Case #3: `Role 1` is restricted to `Study 1`

The user should be able to:

- read `Study 1`
- create new Center
- read, update, destroy `Center 1`
- read `Patient 1`