Convert transfer syntax if requested by query
We disabled the transfer syntax conversions earlier due to performance
reasons. Now we found out, that there are viewers that cannot read
certain transfer syntaxes (e.g. the Weasis Web Viewer, which does not
have JPEG2000 support). These viewers request a certain transfer
syntax.

With this commits, we only convert transfer syntaxes if the request
specifies a transfer syntax explicitly, otherwise we just return the
image as it is.