Implement migration of required series presets
The expected behaviour is:

When ever a visit type changes and required series are specified for
that visit type, the respective required series candidates should be
created.

When migrating from existing data, we have to create required series
versions and records for the respective configurations in history.

This migration creates required series and updates the presets
according to study configuration.