Allow upload of unparsable DICOM files (Fixes ERICA-6)
Upload of DICOM files with embedded PDF content does not work due to
missing pixel data. The upload happens, but the user gets an error
message.

This change introduces separation of warnings and errors in the image
upload component. Parsing errors are mere warnings displayed in orange
and upload errors are errors displayed in red.