Move configuration consolidation after locking a configuration
Before with each upload the configuration was being consolidated. If
the configuration was locked nothing happened. If it was not locked
and a new config was uploaded all required series and visit types were
changed according to configuration.

Now the consolidation takes place when an authorized user locks the
configuration.