Remove `letter_opener` in favor of `mailcatcher`
Before mails were catched by `letter_opener` which is a special mail
delivery method and route within the dev and test rails app. This was
removed in favor of `mailcatcher`.

Now mails are sent via SMTP to a dockerized `mailcatcher` service
linked to the `app` and `test` container, in order to extract mail
testing from the rails app itself.

This way we are not dependent on a special delivery method that may
get in the way with sidekiq and testing.